[nanobench] Disable warmup when explicit loop count is requested
Timings are already disabled when not using --loops 0 (default), so
this ensures that each matched benchmark is run exactly the number of
loops requested. Previously it would run loops up to 1s, which makes
it difficult to collect logs, etc. when debugging or comparing
executions.

Change-Id: I3302461d7fcf4d6b67e47b9bd14ec8b3985c3708
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/998137
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Kaylee Lubick <kjlubick@google.com>