[gn] Remove msvc env setting
SetEnv.cmd has long been removed but kept alive by Skia packaging of
MSVC. Older MSVC compilers and Windows SDKs relied on a large number of
environment variables to be set when cross compiling, but this should no
longer be necessary.

In addition, the MSVC ARM toolchains are now complete and no longer
require the host toolchain to be on the path.

These improvements mean that it is no longer necessary to use `cmd.exe
/c` as a "compiler wrapper" to set up the environment. Because it is not
generally possible to properly escape such command lines this has caused
issues with any arguments which would require proper escaping such as
paths with spaces and special characters.

This change drops support for building Skia for x86 or ARM64 with MSVC
from before VS2022, with the benefit of building for x86 or ARM64
without extemely odd, undocumented, and limiting restrictions.

Also add support for detecting MSVC 2022 before looking for 2019 or
2017.

Bug: 420271794, 376381591
Change-Id: I8591636aec31095b13db6dc1a376abfa23199ebd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/999356
Reviewed-by: Kaylee Lubick <kjlubick@google.com>