[graphite] Separate HSL and regular color matrix filter shaders
The HSL and RGB matrix color filters are rarely used interchangeably,
and already had separate PrecompileColorFilter APIs. Having the shader
code include logic for the HSL <-> RGB conversion added a lot of
instruction/register overhead that hurt performance even when the branch
was entirely uniform.

Adds a GM that leverages the RGB matrix color filter in multiple large
screen draws to highlight its cost. This is representative of pipelines
that are created in the Android swipe+blur user interaction when there
is an HDR video PiP widget on screen.

The splitting improves GPU time from 10ms to 6ms on my M1. There are
similar improvements in GPU time in Vulkan on my linux machine and on
pixel devices.

Bug: b/419551614
Change-Id: Id7dcc791fb96a43072a80a35b7b7f82a2b730c76
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/994867
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>