Add bounds check to onDrawPoints for ganesh Device
This fuzzer bug was due to a change to `SkCanvas::onDrawPoints` made in
this CL: https://skia-review.googlesource.com/c/skia/+/995176. The
change made it such that we didn't compute bounds when drawing points,
but we actually need to in order to check `internalQuickReject` by the
canvas. Our `localToDevice` matrix had inf values, so we do not want to
add an OpsTask, but when we do not check `internalQuickReject`, the task
is added and eventually flushed (which is where we see the failed ASSERT
in the stack trace).

Since the cpu backend already checks for valid bounds, as well as
graphite (in `Device::drawGeometry()`), we don't want to check the
bounds twice so we can do it in `ganesh::Device::drawPoints` to avoid
this.

Bug: 419100000
Change-Id: I0ec6b2ab884239021b8b45f6a86c5e3aba285013
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/999717
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>