More uses for DirectBlit
- extend rasterpipeline blitter
- use in hairlines (2-3x faster) and hairpoints (3x+ faster)
- add a8 and gray8 to nanobench

The goal is to greatly speed up drawing single-points when the
src color can be written directly to the dst (i.e. no blending).
Hence the expansion of the benchmark to include src-over and src
modes.

Hairpoints and Hairlines both compute their values one point
at a time, so both are candidates for this faster entry-point.

Change-Id: I8b70958df40e40327755d4d1da0351098018f365
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/997499
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Mike Reed <mike@reedtribe.org>
Reviewed-by: Daniel Dilan <danieldilan@google.com>