use SkSpan in SkTypeface API
Will update where the guard is defined after the SkFont CL lands.

There are virtuals that should be updated to use Span as well...
but do we have clients that create subclasses? If so, that migration
will be tricky. Hence this CL doesn't attempt to update those (yet).

Change-Id: I9d3e45623d309c75a361f1b73d6fa20dcbd1e806
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1001356
Reviewed-by: Ben Wagner <bungeman@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Mike Reed <mike@reedtribe.org>