Reapply "use SkSpan in SkTypeface API"
The prev version used 1<<30 for a max tabletag count. This was too
large for a 32bit machine (failed an assert). In hindsight, its
larger than needed given what we know about opentype fonts anyway.

New value is 1<<16 - 1, which should be fine on a 32bit machine,
and large enough for fonts. Also, renamed the 'new' version to
readTableTags(span) so has not to hide the legacy getTableTags(ptr)
so we can test both paths for now.

This reverts commit f28e9e508fcbd52ff80c0612e1b1ff8867467afb.

Change-Id: I9f31de03a07fcf857b3aaed010651a1f9d428138
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1002458
Reviewed-by: Ben Wagner <bungeman@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Mike Reed <mike@reedtribe.org>