SkColorSpaceXformSteps: Start to add support for HLG and PQ
Add support for the new HLG and PQ transfer functions to
SkColorSpaceXformSteps and SkColorSpaceXformSteps::apply.

Add parameters for the HLG opto-optical transfer function, since they do
not exist in SkColorSpaceXformSteps yet.

In the constructor of SkColorSpaceXformSteps, scale the gamut transform
matrix as needed by the peak luminance and HDR reference white
luminance.

This is not a complete implementation (it only adds the parameters and
tests the lowest performance path). Support for these parameters will be
needed in the transform operations of the software, ganesh, and graphite
paths.

Bug: skia:420956739
Change-Id: I6b6ab5763921151ad8bee8cb25553c48ec986143
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1002536
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Christopher Cameron <ccameron@google.com>
Reviewed-by: Christopher Cameron <ccameron@google.com>