Implement drawCoverageMask for SkBitmapDevice
Removes flag that stops the compilation of the functions drawDevMask()
and drawBitmapAsMask() from SkDraw class, and uses it to draw an A8
bitmap as coverage. Also adds a parameter to them to indicate the
transform for the paint, which is different than SkDraw.fCTM.

Not currently load bearing (until useCoverageForMaskFilter() set to true
for SkBitmapDevice), when enabled changes behavior of raster backend for
tablemaskfilter, and some blend modes for embossmaskfilter.

Bug: 304823483
Change-Id: I3a157ce9d7b1c72f2a5ad3acdfbe859c5b2481cc
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/997356
Auto-Submit: Daniel Dilan <danieldilan@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>