Add end to end colorspace test
Generated png's and jpeg's tagged with ICC profiles with DCI-P3 and
AdobeRGB, as well as the same images transformed to the other
colorspace. The code to generate the images is not compiled. The
images are decoded into different colorspaces and tested against the
generated golden images.

Change-Id: I4a99b919e0abe168da3d21da68563b4d532e0970
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1005488
Commit-Queue: Daniel Dilan <danieldilan@google.com>
Reviewed-by: Christopher Cameron <ccameron@google.com>
Auto-Submit: Daniel Dilan <danieldilan@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>