[vulkan] Use std::vector<uint32_t> for SPIR-V
Instead of std::string. SPIR-V is defined in 32-bit words and includes
lots of zeros. Use of std::string (due to other backends) is awkward.
This change makes the "native shader" be a type that can either contain
text or binary. As a bonus, it makes it less error-prone to mix up
shader input and output std::strings.

As a small drive-by, the skia_print_native_shaders feature now works
with SPIR-V as well.

Bug: b/390458117
Change-Id: I65326b043e9634651071b66403906a2295fbaacc
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1006716
Commit-Queue: Shahbaz Youssefi <syoussefi@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>