[graphite] Visit clip stack before choosing renderer
visitClipStackForDraw does not take outsetBoundsForAA anymore, so there
isn't a dependency on the Renderer anymore. After visiting the clip
stack, and after the Renderer is chosen, the Clip's drawBounds and
transformedShapeBounds are adjusted to account for the AA outset.

This may influence the usage bounds of affecting clip elements and
whether or not the scissor is applied, and the ordering used between
draws, but it isn't necessary for deciding whether or not a logical clip
element affects a shape (which is done based on the original geometry)

Bug: b/424504941
Change-Id: I449a298d6f6d3f9a79a24bf4f735b0ac18005aec
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1007136
Reviewed-by: Thomas Smith <thomsmit@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>