move pathops into a module
Goal is to make it clear what parts of the source base are
fundamental to core, and which are optional. Modules captures that
notion pretty well.

Note: The printing backends (PDF, XPS) are possibly candidates to move
into modules as well -- they are clearly optional (to the core) just
like SVG or Shaper.

Change-Id: I9d5d51b380a290c419295da8277590607d007192
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/999760
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Florin Malita <fmalita@google.com>