[graphite] De-register gen ID listeners added by ProxyCache
Gen ID listeners are remembered in the cache so that if the cache
entry is removed due to timeout/purge, the listeners can be marked
to deregister from the bitmap they were added to.

This allows ProxyCache to go back to adding listeners to immutable
bitmaps (since the listener still reports destruction), fixing the
leak reported in b/370696605. But because the listeners are now
deregistered, we also avoid the leak that https://skia-review.googlesource.com/c/skia/+/882256
was working around.

While we could have a fast path for clients to say this SkBitmap
is immutable and lives forever, that puts a lot of responsibility on
the caller. Now we may add/remove/add/etc. a listener on the dither
LUT when cleaning unused proxies, but that is pretty minimal for
the consistency in handling bitmap lifetimes.

Bug: b/370696605
Bug: chromium:337980338
Change-Id: Ie9dad7fbd3d1fafd3373cec23163773e9d82948b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1008416
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Thomas Smith <thomsmit@google.com>