[graphite] Generate blurred rrect profile on the GPU
Adds a findOrCreateCachedProxy that uses a generator function that
is provided a Recorder and can return a Graphite texture Image.

The ProxyCache handles making sure the image isn't actually scratch,
and getting the tasks added to the root task list of the Recorder.
Updates AnalyticBlurMask to use a blur image filter applied to the
canonical 9-patch rrect to skip calling CreateRRectProfile.

Bug: chromium:420959038
Bug: 338032240
Change-Id: Ie91d3faef7d529753d369647664d84fc12a5aa5a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1008419
Reviewed-by: Thomas Smith <thomsmit@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>