Introducing SkPathRaw
The idea is to factor out the 'content' of a SkPath, but remove
any/all of the particulars of ownership or allocation.

In this 'raw' form, we should be able to do all of the functional
operations: iterate, produces edges, vertices, query, stroke, etc.

Plan is to (in future CLs) refactor SkPath and SkPathBuilder to
return a raw, allowing lower-level code (like scan-converters,
strokers, path-measuring) to accept a Raw instead of requiring a
Path.

This refactoring will in-turn allow some clients to bypass ever
creating (and allocating memory) an actual path, and instead just
create a raw (e.g. on the stack) and still perform the operation.
(e.g. SkBitmapDevice::drawCircle might create a raw containing the
necessary conics, draw it, and never have to create a SkPath).

Change-Id: I905b65aa2e2b032d1d9002db58155ed8ed19a5d2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1004157
Commit-Queue: Mike Reed <mike@reedtribe.org>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>