[graphite] Add SK_TRACE_GRAPHITE_PIPELINE_USE build variant, expand tracing
Recorder::snap and QueueManager::addRecording are now using _ALWAYS, so
Graphite's heavier CPU work will be traced analogous to the events that
Ganesh always traces in Android.

This also adds the SK_TRACE_GRAPHITE_PIPELINE_USE build flag that will
sum up the draw bounds pixel area for each draw per pipeline and then
trace each pipeline label and its approximate rasterization area in
DrawPass::prepareResources. The labels are saved on the
GraphicsPipelines, so the main overhead in tracing this is that the
strings are somewhat long, but it's per-DrawPass per-pipeline so not
coupled to the number of draws issued.

This adds INSTANT_ALWAYS variants that work the same as EVENT_ALWAYS but
for just reporting values w/o a time span.

Change-Id: I3739a8fef1d39dc42984e9d0ec340b7fe9234429
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1009824
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Noelle Scobie <nscobie@google.com>