Implement specialized A8 blur to SkBlurEngine from SkMaskBlurFilter
This CL adds a blur algorithm RasterA8BlurAlgorithm, which is the same
idea as Raster8888BlurAlgorithm, but to be used for a8 colortype. The
other difference is that this uses the new Pass extension, A8Pass class.
This is also the same as ThreeBoxAppoxPass, but the templated
implementation was getting messy because of the lsx support (if we could
template this we could get rid of A8pass and combine it with the
ThreeBoxApproxPass and then make just one Algorithm class too).

Most of the blurring work for the Algorithm has been put into a static
function eval_blur_passes(), to be used by both the rgba8888 and a8
Algorithm implementations. The correct colortypes are asserted in the
Algorithm subclasses, which are the callers of the templated functions,
and is important because the current implementation requires 8bit
color channels.

For small blurs, we can now use GaussianPass, which is a true gaussian
blur with inspiration from this CL [1]. Since it applies the gaussian
kernel directly, it has a much smaller max sigma compared to the
box-blur approximation, but is accurate for the small sigmas. It is much
faster than the raster pipeline approach. Thus this is templated and can
be used in both the rgba8888 and a8 algorithms.

Future work may be to implement the small blur algorithm to fit into
this algorithm->pass design for A8 that is implemented in
SkMaskBlurFilter.cpp and may or may not be faster than this
GaussianPass, as well as the templating mentioned above.

[1] https://skia-review.googlesource.com/c/skia/+/1008177

Bug: 40042849
Change-Id: I7c5af9841ba4dc75bcc016eec6a294883329ce66
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1002518
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Daniel Dilan <danieldilan@google.com>