[graphite] Treat linear/encode as sRGBish for fast path
The ColorSpaceTransformSRGB shader block was requiring that both the
src and the dst transfer functions be sRGBish and that the steps
had `linear` and `encode` set to true. However, if doing a colorspace
for the toLinearSrgb()/fromLinearSrgb() intrinsics (or manually writing
linear data, etc), you could very well have one side be sRGB-ish and
the other be linear.

The linear/encode identity functions are representable in this same
code path, so it relaxes them to let linear/encode be false in that
case.

Change-Id: Id14e5a0586912bbb66887aa0b7792ef9f880d43e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1010457
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>