[graphite] Additional Android-specific PaintOptions
In addition to adding some new PaintOptions this CL also upgrades PrecompileSettings to take an SkSpan of RenderPassProperties. This allows us to remove some of the duplication in the kPrecompileCases array.

This CL also removes the experimental "deduce_settings_from_label" code. This was an attempt to automate the creation of the PaintOptions from the Pipeline labels.

Bug: b/390186667
Change-Id: Ia134ded68c7089b906e38ea97a551d5dc5a3ce58
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1011816
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>