[graphite] Use correct working context for precompile's withWorkingColorSpace/format
The shader just needed to use the existing `workingContext` local.
This was unused, but maybe being masked by the `[&]` lambda captures so
we weren't getting any warnings that would have caught the issue sooner.

The working format color filter is updated to compute the working
colorspace and alpha type, and create the working context to exactly
match the runtime logic. The working context is then also used as
the KeyContext for the child.

Change-Id: I0d8ba8ad155615e350b7b2e17ab546fbf3890781
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1012676
Commit-Queue: Robert Phillips <robertphillips@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Auto-Submit: Michael Ludwig <michaelludwig@google.com>