Allow fiddle_examples in shared builds
Most of Skia's tests are "big ball of mud" tests which test the
internals of Skia and need access to all the symbols. These do not work
in a shared build. However, the point of fiddle_examples is to
demonstrate the use of only public API. Allow this to be built in a
shared build configuration. This allows detection of use of unexported
API in the examples, in particular when it was thought an API was
exported but the defintition actually was not exported.

Change-Id: Iab0550bb2b5bd03986bf309fb6b6ab51ee3c1632
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1013976
Reviewed-by: Florin Malita <fmalita@google.com>
Auto-Submit: Ben Wagner <bungeman@google.com>
Commit-Queue: Florin Malita <fmalita@google.com>