[graphite] Add a PaintOptions Builder
This will eliminate a lot of the boiler plate of PaintOptions construction.

For this to work we would need to copy the contents of PaintOptionsBuilder.h/.cpp directly into Chrome and Android. I don't think it warrants being part of the Skia API.

Ultimately, I would like to not even have the trampoline functions (e.g., SolidSrcover).

Bug: b/390186667
Change-Id: I9eaa553a94285cebe334da58f0f0d99ac4e1c0bd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1012416
Commit-Queue: Robert Phillips <robertphillips@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Noelle Scobie <nscobie@google.com>