Make SkImage->makeSubset take an SkRecorder* instead of a graphite recorder
With this change, there are no more references to a graphite recorder
in SkImage. There are a few APIs that still reference GrDirectContext
(and don't have an option for SkRecorder) and I've decided to leave
those be for now:
 - makeNonTextureImage
 - makeRasterImage
 - readPixels
These do a readback and won't work on Graphite nor async CPU, so I
think having them take an SkRecorder is misleading. We can re-visit.

Change-Id: If9ee91500b84fbd83107684dcf90bd358ab80317
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/999656
Reviewed-by: Jorge Betancourt <jmbetancourt@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>