[graphite] Unify clear/discard load op handling inside drawGeometry
The logic for determining if a paint depends on the dst color is
consolidated to just support a `PaintParams`, there's no need to invoke
it with an SkPaint anymore (which had only been because drawPaint did it
early).

Now clearing and discarding is handled in `drawGeometry()` after the
clip stack has had a chance to analyze the geometry. Since the clip
stack converts [r]rects and quads that cover the target into flood
fills, it now means any simple convex shape (currently intentionally
excluding paths from this analysis) can become clear load ops or discard
prior work.

Before this CL it only happened for drawPaint calls, so calling
drawImageRect or drawRect that covered the target wouldn't give you the
benefits.

Bug: b/424509730
Change-Id: Idbfadcc5388e871e028c5a160e79687d54cccf6b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1016036
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>