[graphite] Refactor visitClipStack to use DrawShape helper
This shouldn't really be a functional change. It re-arranges the bounds
calculations for the draw into a helper object called DrawShape that
knows how to represent itself as a TransformedShape (so analogous to
RawElement for clips).

This will make it easier to keep track of modifications to the draw and
its bounds when the clip elements are allowed to modify the geometry.
Even before that, I think it makes the visitClipStackForDraw function
less daunting.

Bug: b/424508534
Change-Id: I28b38bb35edcc6d508454ff95a296f67540a40df
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1015996
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>