[graphite] Add debugging members for ResourceCache usage
Adds fields to gauge how fast resources are going through the return
queue, the purgeable queue, the proxy cache, and the upload buffer
manager. Hopefully this will help narrow down the cadence of when
resources are being requested and returned, and what type of resource
that is being accumulated to pathologic quantities.

Also updates the SkTMultiMap to track the size of values stored per
key to help identify distribution there.

Bug: b/407062399
Change-Id: Iae4aacec217ad12c1f290c46ad88081d0b6b4277
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1017376
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>