[graphite] Allow ClipStack to modify draw geometry
This allows the ClipStack to replace a rectangle, rounded rectangle,
fully-antialiased PerEdgeAAQuad that is rectangular, and a flood fill
with a [rounded]rectangular clip element's geometry. It can also
geometrically intersect the clip elements' geometries against draw
geometry of the same class. This can cause an edge-AA rect or flood fill
to change to a rrect draw, and a round rect draw to change to a flood
fill. Since these renderers are often used in the same family for
precompile, this hopefully doesn't change the complexity of identifying
what to precompile. But that is why the ability to swap out to a flood
fill is not extended to arbitrary convex paths that cover the clip, or
to replace a covering rectangular draw with a clip path.

This allows more clips to be handled without MSAA, and in particular
should almost entirely eliminate the possibility of a clipRRect +
drawPaint or drawRect that covers it would trigger MSAA.

This does not yet take advantage of clear/discard load ops, although it
sets the stage for that as Device can now just check if the final
geometry is an unclipped flood fill, regardless of what had come before
hand.

Bug: b/424508534
Change-Id: I149b99f9a6e667dd93ed4ab27c26c11d556fc7e8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1015997
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>