[graphite] Snap scissor rects to 4px boundaries
Before this change, the clip stack would set the scissor per draw to the
outer pixel bounds of the stack. This was tight and minimized the amount
of rasterization of draws, and minimized the amount of inverse fill
required for the intersect clip depth draws. However, with analytic
coverage outsetting the draw bounds, the scissor was almost always
considered required. When clipping was involved, it would significantly
impact potential batching because everything required a unique scissor.

With this change, the scissor is snapped to a 4px boundary that contains
the original scissor. This avoids requiring a scissor just because of AA
outsets and when the scissor *is* still required, it avoids some
thrashing because we've quantized the possible values. It adds
additional logic to tracking the usage bounds and scissor for clip draws
to account for this potential increase in pixels that have to be covered
by the depth-only clip draws, and adds a heuristic to prefer just
rasterizing a larger clip area to avoid setting a scissor. For now the
heuristic is based on area change from a 45 degree rotation (e.g. the
scenario you'd get by clipping to a rectangle and a rotation of that
same rectangle to form an octagon).

Clip perf summary -----------------

By the end of this change stack, (starting with
https://skia-review.googlesource.com/c/skia/+/1009396), some motionmark
scenes GPU work improves?

motionmarksuitsclip SKP goes from ~7200 draw calls, ~16k commands, and a
GPU time of 3.1ms on an M3 to ~1200 draw calls, ~3300 commands and a GPU
time of 2.7ms. motionmarksuits SKP goes from ~5100 draw calls, ~12k
commands and a GPU time of 3.8ms to 2300 draw calls, ~6k commands and a
similar GPU time.

On nanobench, grmtl suits is unchanged (~5.4ms) but suitsclip goes from
4.55ms to 4ms. The reduction in commands more significantly improves
grdawn, with suits going from 6.41 to 6.1ms and suitsclip going from
5.14ms to 4.3ms. The other motionmark tests (that don't really trigger
clipping, so are a good measure for base overhead) don't really change.

Bug: b/424510276
Change-Id: I36c0082ffe8699c59a9e01f98e3335e0bb6fffa3
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1015960
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>