[graphite] Only add RenderStep coverage SkSL if step reports coverage
Previously, the fragmentCoverageSkSL() function was being added to the
shader code if anything triggered analytic coverage. This was mostly
hidden because when the RenderStep wasn't the source of the analytic
coverage, it was just returning an empty string.

BitmapTextRenderStep is in a unique position because it either has
fragment coverage OR primitive colors depending on its mode of
operation. It doesn't change what it returns from the sksl functions,
but does change which flags it reports.

With the prior scissor changes, we were switching to use an analytic
clip. This caused an RGBA-color text renderstep (which is primitive
color, not single/LCD coverage) to also have analytic coverage. The
fragment coverage function was still adding the sksl for sampling
the text atlas (which actually held colors), but was then
interpreting it as the base analytic coverage to be combined with
the analytic clip.

Change-Id: Icfb72b888544eccf0b02e274e498ffaaf3a0f6cd
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1021476
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>