[pdf] Fix zero extended glyph to code point maps
Prior to this change any extended glyph to code point maps (mappings
from a single glyph to a non-default set of code points) emitted both a
mapping to U+0000 and the actual mapping. Remove this extra incorrect
mapping and update the existing test to reflect that all used glyphs
will be marked as being in the subset to test this change.

Bug: 40043392
Change-Id: I22c4eba2064d22ea946478093a05dbd8b0dab6c2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1022376
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>