[graphite] Fix draw/clip contains priority
This should fix the perf regressions introduced from the inner bounds CL
shown in
https://skia-perf.luci.app/e/?begin=1750917470&enable_chart_tooltip=true&end=1753099846&keys=Xe26b184787f69bc459d2de2d8f2470da&plotSummary=true&show_google_plot=true&use_titles=true&xbaroffset=75803&request_type=0

With the inner bounds change, when a clip and a draw both contained each
other (e.g. they were the same rectangle), the b.contains(a) check would
now pass with the early inner bounds test and we'd return kReplacesDraw
instead of kNone for the influence. But before
https://skia-review.googlesource.com/c/skia/+/1015997 landed, this was
handled as if the clip intersected the shape now.

While the clip+draw modifications resolved a lot of the regressions by
intersecting the clip directly, it's better to just prefer treating
these situations as unclipped.

Bug: b/424506312
Change-Id: I9932902a8296fe004eb98ef586bbfc59a4d5bb3e
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1023296
Reviewed-by: Thomas Smith <thomsmit@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>