[graphite] Fix paint_depends_on_dst logic
Fixes xfermodes3 GM, where a clear blend mode is the same as a
transparent clear, regardless of the paint.

Also fixes logic for analyzing the primitive blender, since that doesn't
impact the dst blend, it just impacts whether or not we know the
src is opaque.

Re-arranges the logic to only calculate whether or not the src is
opaque if we need it (for src-over and now, dst-out).

Change-Id: Ieca75b9dba34fe91394c10461317f3ba3352a399
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1023376
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>