[pdf] Avoid inverse in drawPaint
Similar to "[graphite] Guard clip/draw intersections with precision
check" [0] the PDF code relied on an inverse matrix for placing
drawPaint (which is used by SkCanvas::clear). The PDF code used this
inverse to compute a rectangle in CTM units which would cover the
current clip which is specified in device units. However, with a large
translation the lack of precision in float space can lead to an
incorrect rectangle. Instead, just draw the rectangle that covers the
current clip without the CTM and forward apply the CTM to the paint's
shader.

This fixes the PDF output for the cliplargerect gm.

[0] https://skia.googlesource.com/skia.git/+/32a2a9b920130866333232a29eb0311e0ecef895

Change-Id: I9e053d294305d1f8a9c76a8b47abeb50f503c00f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1025836
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>