[graphite] Fix bug(s) in PipelineKey serialization
The first erroneous check was intended to ensure that the next sample we were going to make into keyData was valid.

The second one was to verify that the data could've fit in keyData.

I believe they both slipped through since any block that stores a Sampler isn't going to be the last one in the key.

Change-Id: I9dd1303d493940da862027130fb8a29c0521231b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1026756
Commit-Queue: Robert Phillips <robertphillips@google.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>